Changes to fix the issues and add Transactional things

+++ lab2/src/main/java/edu/sjsu/cmpe275/lab2/controller/EmployeeController.java
@@ -6,6 +6,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
+import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
@@ -14,6 +15,7 @@ import org.springframework.web.bind.annotation.RestController;
 
 import edu.sjsu.cmpe275.lab2.model.Address;
 import edu.sjsu.cmpe275.lab2.model.Employee;
+import edu.sjsu.cmpe275.lab2.model.Employer;
 import edu.sjsu.cmpe275.lab2.service.EmployeeService;
 import edu.sjsu.cmpe275.lab2.service.EmployerService;
 
@@ -30,15 +32,18 @@ public class EmployeeController {
 	public ResponseEntity<Object> getAllEmployees() {
 		List<Employee> emps = employeeService.getAllEmployees();
 		return ResponseEntity.ok(emps);
-		//return emps;
 	}
 	
 	@RequestMapping("/employee/{id}")
 	public ResponseEntity<Object> getEmployee(@PathVariable long id) {
-		Employee emp = employeeService.getEmployee(id);
-		return ResponseEntity.ok(emp);
+		Employee emp =  employeeService.getEmployee(id);
+		if (emp!=null)
+			return ResponseEntity.ok(emp);
+		else
+			return ResponseEntity.status(HttpStatus.NOT_FOUND).build(); 
 	}
 	
+	@Transactional
 	@RequestMapping(method=RequestMethod.POST, value="/employee", produces = { "application/json", "application/xml" })
 	public ResponseEntity<Object> addEmployee(@RequestParam String name
             , @RequestParam String email
@@ -55,13 +60,27 @@ public class EmployeeController {
 		emp.setEmail(email);
 		emp.setTitle(title);
 		emp.setAddress(new Address(street, city, state, zip));
-		emp.setEmployer(employerService.getEmployer(employerId));
-		emp.setManager(employeeService.getEmployee(managerId));
-		//emp.setEmployer();
+		Employer tempEmp = employerService.getEmployer(employerId);
+		if (tempEmp != null) {
+			emp.setEmployer(tempEmp);
+		}
+		else {
+			return ResponseEntity.status(HttpStatus.NOT_FOUND).build();
+		}
+		Employee tempMang = employeeService.getEmployee(managerId);
+		if (tempMang != null) {
+			if (tempMang.getEmployer().getId() == employerId) {
+				emp.setManager(tempMang);
+			}
+			else {
+				return ResponseEntity.status(HttpStatus.BAD_REQUEST).build();
+			}
+		}
 		employeeService.addEmployee(emp);
 		return ResponseEntity.ok(emp);
 	}
 	
+	@Transactional
 	@RequestMapping(method=RequestMethod.PUT, value="/employee/{id}", produces = { "application/json", "application/xml" })
 	public ResponseEntity<Object> updateEmployee(@PathVariable long id, @RequestParam String name
             , @RequestParam String email
@@ -82,17 +101,48 @@ public class EmployeeController {
 			emp.setEmail(email);
 		if (title != null)
 			emp.setTitle(title);
-		if (employerId != null)
+		if (employerId != null) {
+			if (emp.getEmployer().getId() != employerId) {
+				Employee thisManager = emp.getManager();
+				List<Employee> thisReports = emp.getReports();
+				for (Employee e : thisReports) {
+					e.setManager(thisManager);
+				}
+			}
 			emp.setEmployer(employerService.getEmployer(employerId));
-		if (managerId != null)
-			emp.setManager(employeeService.getEmployee(managerId));
+		}
+		if (managerId != null) {
+			Employee tempMang = employeeService.getEmployee(managerId);
+			if (tempMang != null) {
+				if (tempMang.getEmployer().getId() == employerId) {
+					emp.setManager(tempMang);
+				}
+				else {
+					return ResponseEntity.status(HttpStatus.BAD_REQUEST).build();
+				}
+			}
+			else {
+				return ResponseEntity.status(HttpStatus.NOT_FOUND).build();
+			}
+		}
 		employeeService.updateEmployee(emp);
 		return ResponseEntity.ok(emp);
 	}
 	
+	@Transactional
 	@RequestMapping(method=RequestMethod.DELETE, value="/employee/{id}", produces = { "application/json", "application/xml" })
 	public ResponseEntity<Object> deleteEmployer(@PathVariable long id) {
 		try {
+			if (employeeService.getEmployee(id).getReports().size() > 0)
+				return ResponseEntity.status(HttpStatus.BAD_REQUEST).build();
+			Employee emp1 = employeeService.getEmployee(id);
+			List<Employee> collab = emp1.getCollaborators();
+			for (Employee e: collab) {
+				e.getCollaborators().remove(emp1);
+				employeeService.updateEmployee(e);
+			}
+			emp1.getCollaborators().clear();
+			employeeService.updateEmployee(emp1);
 			employeeService.deleteEmployee(id);
 		}
 		catch(Exception e) {
